+++ app/models/concerns/active_fields/field_concern.rb
@@ -12,14 +12,22 @@ module ActiveFields
         inverse_of: :active_field,
         dependent: :destroy
 
-      scope :for, ->(customizable_type) { where(customizable_type: customizable_type) }
+      # Returns active fields for the given customizable type, including both
+      # scoped fields (when scope parameter is provided) and global fields (where scope is nil).
+      scope :for, ->(customizable_type, scope: nil) do
+        # Global fields are available to all records of the given customizable type.
+        scopes = [scope, nil].uniq
+        where(customizable_type: customizable_type, scope: scopes)
+      end
 
       validates :type, presence: true
-      validates :name, presence: true, uniqueness: { scope: :customizable_type }
+      validates :name, presence: true
+      validate :validate_name_uniqueness
       validate :validate_default_value
       validate :validate_customizable_model_allows_type
 
       after_initialize :set_defaults
+      before_validation :set_scope
     end
 
     class_methods do
@@ -88,7 +96,7 @@ module ActiveFields
 
     # Returns customizable types that allow this field type.
     #
-    # Notes:
+    # NOTE:
     # - The customizable model must be loaded to appear in this list.
     #   Relationships between customizable models and field types are established in the `has_active_fields` method,
     #   which is typically called within the customizable model.
@@ -102,6 +110,26 @@ module ActiveFields
     end
 
     private
+
+    # NOTE: The uniqueness constraint in the DB does not fully enforce this validation:
+    # Records where scope is NULL do not prevent the existence of records with the same
+    # [name, customizable_type] but with a non-NULL scope, and vice versa.
+    def validate_name_uniqueness
+      base_scope =
+        ActiveFields.config.field_base_class.excluding(self).where(customizable_type: customizable_type, name: name)
+
+      if scope.nil?
+        if base_scope.exists?
+          errors.add(:name, :taken)
+          return false
+        end
+      elsif base_scope.exists?(scope: [scope, nil])
+        errors.add(:name, :taken)
+        return false
+      end
+
+      true
+    end
 
     def validate_default_value
       validator = value_validator
@@ -127,5 +155,12 @@ module ActiveFields
     end
 
     def set_defaults; end
+
+    # Forces scope to nil if the customizable model does not have a scope method.
+    def set_scope
+      return if customizable_model.nil?
+
+      self.scope = nil if customizable_model.active_fields_scope_method.nil?
+    end
   end
 end