+++ spec/dummy/app/controllers/authors_controller.rb
@@ -49,7 +49,7 @@ class AuthorsController < ApplicationController
       author: [
         :name,
         :group_id,
-        active_fields_attributes: [:name, :value, :_destroy, value: []],
+        active_fields_attributes: [[:name, :value, :_destroy, value: []]],
       ],
     )
     permitted_params[:active_fields_attributes]&.each do |_index, value_attrs|