+++ spec/dummy/app/javascript/controllers/active_field_form_controller.js
@@ -1,0 +1,19 @@
+import { Controller } from "@hotwired/stimulus"
+
+export default class extends Controller {
+  static targets = [
+    "scopeInput",
+    "disableScopeInput",
+  ]
+
+  toggleScopeInput(e) {
+    e.preventDefault()
+
+    if (this.disableScopeInputTarget.checked) { 
+      this.scopeInputTarget.disabled = true
+      this.scopeInputTarget.value = ""
+    } else {
+      this.scopeInputTarget.disabled = false
+    }
+  }
+}