+++ spec/dummy/app/policies/active_fields/field/text_policy.rb
@@ -4,7 +4,7 @@ module ActiveFields
   module Field
     class TextPolicy < ApplicationPolicy
       def permitted_attributes_for_create
-        %i[customizable_type name min_length max_length default_value]
+        %i[customizable_type name scope min_length max_length default_value]
       end
 
       def permitted_attributes_for_update