+++ spec/dummy/db/migrate/20240229220000_create_dummy_tables.rb
@@ -19,5 +19,11 @@ class CreateDummyTables < ActiveRecord::Migration[7.1]
       t.references :author
       t.timestamps
     end
+
+    create_table :users do |t|
+      t.string :email
+      t.string :tenant_id
+      t.timestamps
+    end
   end
 end